SAP_Solar_Project/requirements.txt
flask
flask-sqlalchemy
flask-login
pandas
matplotlib
openpyxl
pdfplumber
